% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gw_collin_plot.R
\name{gw_collin_plot}
\alias{gw_collin_plot}
\title{Mapping of multicollinearity}
\usage{
gw_collin_plot(obj, metric = "vif")
}
\arguments{
\item{obj}{A pargwr object fitted with gwr_par}

\item{metric}{String. The multicollinearity metric to be plotted (either 'vif' or 'corr')}

\item{plot}{Logical. Plot a map if TRUE, return a ggplot2 object if FALSE}
}
\description{
Creates a map of multicollinearity indices (local correlation and local VIF) computed when the parameter 'collinearity' of the gwr_par function is set to TRUE.
}
