% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwr.internal.R
\name{gwr.internal}
\alias{gwr.internal}
\title{Internal function. Computes the local LM for a given point.}
\usage{
gwr.internal(x, y, yhat, cell, coords, bandwidth, weights = NULL, kernel,
  longlat, adapt, se.fit)
}
\description{
Internal function. Computes the local LM for a given point.
}
